+++ src/d3.easygraph.heatmap.js
@@ -19,6 +19,7 @@ d3.easygraph.heatmap = function(config) {
   // wins (config is folded onto graph before familyDefaults is even consulted).
   return d3.easygraph._build(config, {
     _chartType:   'heatmap',
+    _dataShape:   'grid',
     colorPalette: 'Diverging.RdBu.reversed'
   }, function(graph) {
     function render(data) {